Move ExtractFunctionName tests
